py/requirements.txt
streamlit
streamlit-authenticator
pandas
pyyaml
pillow
openpyxl
